Bruno/add prettier to sol files (#60)
* Added prettier and prettified .sol files

* Prettified a file that was left

* Added check for .sol prettier

* Joined script into a single command

+++ packages/smart/contracts/balancer/BFactory.sol
@@ -19,28 +19,17 @@ pragma solidity 0.7.6;
 import "./BPool.sol";
 
 contract BFactory is BBronze {
-    event LOG_NEW_POOL(
-        address indexed caller,
-        address indexed pool
-    );
+    event LOG_NEW_POOL(address indexed caller, address indexed pool);
 
-    event LOG_BLABS(
-        address indexed caller,
-        address indexed blabs
-    );
+    event LOG_BLABS(address indexed caller, address indexed blabs);
 
-    mapping(address=>bool) private _isBPool;
+    mapping(address => bool) private _isBPool;
 
-    function isBPool(address b)
-        external view returns (bool)
-    {
+    function isBPool(address b) external view returns (bool) {
         return _isBPool[b];
     }
 
-    function newBPool()
-        external
-        returns (BPool)
-    {
+    function newBPool() external returns (BPool) {
         BPool bpool = new BPool();
         _isBPool[address(bpool)] = true;
         emit LOG_NEW_POOL(msg.sender, address(bpool));
@@ -54,26 +43,19 @@ contract BFactory is BBronze {
         _blabs = msg.sender;
     }
 
-    function getBLabs()
-        external view
-        returns (address)
-    {
+    function getBLabs() external view returns (address) {
         return _blabs;
     }
 
-    function setBLabs(address b)
-        external
-    {
+    function setBLabs(address b) external {
         require(msg.sender == _blabs, "ERR_NOT_BLABS");
         emit LOG_BLABS(msg.sender, b);
         _blabs = b;
     }
 
-    function collect(BPool pool)
-        external
-    {
+    function collect(BPool pool) external {
         require(msg.sender == _blabs, "ERR_NOT_BLABS");
-        uint collected = IERC20Balancer(pool).balanceOf(address(this));
+        uint256 collected = IERC20Balancer(pool).balanceOf(address(this));
         bool xfer = pool.transfer(_blabs, collected);
         require(xfer, "ERR_ERC20_FAILED");
     }
