+++ packages/smart/contracts/balancer/Migrations.sol
@@ -3,7 +3,7 @@ pragma solidity 0.7.6;
 
 contract Migrations {
     address public owner;
-    uint public lastCompletedMigration;
+    uint256 public lastCompletedMigration;
 
     constructor() {
         owner = msg.sender;
@@ -13,7 +13,7 @@ contract Migrations {
         if (msg.sender == owner) _;
     }
 
-    function setCompleted(uint completed) external restricted {
+    function setCompleted(uint256 completed) external restricted {
         lastCompletedMigration = completed;
     }
